Adding LFP extraction

+++ pipeline_utils.py
@@ -47,7 +47,7 @@ def extract_LFP(config_kilosort):
     all_data = np.zeros((int(data.ns/30), data.nc), dtype=np.float32)
     print(all_data.shape)
     all = list(range(data.ns))
-    buffer_size = 30000*4.5
+    buffer_size = int(30000*4.5)
     intervals = all[0: int(data.ns): 1000000]
     print(intervals)
     I = np.zeros(2, dtype=np.int64)
